requirements.txt
python-telegram-bot==20.3
supabase==2.24.0
requests
